pyproject.toml
[project]
name = "zorvyn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.6",
    "fastapi>=0.135.3",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pymysql>=1.1.2",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.4.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
